feat: habilitar C AVX2 (fix overflow i32 + cargo-zigbuild link)
Mudancas:

  - native/distance_avx2.c: int32_t -> int64_t (evita overflow)

  - native/distance_avx2.h: declarar int64_t

  - api/src/dataset.rs: declarar extern como i64

  - api/build.rs: usar cargo:rustc-link-arg para lib path

    (cargo-zigbuild nao pega cargo:rustc-link-lib automaticamente)

O C AVX2 agora eh compilado e linkado no binario.

Isto habilita o uso de instrucoes AVX2 para calcular L2 distance,

que eh mais rapido que a implementacao Rust pura.

+++ api/src/dataset.rs
@@ -302,7 +302,7 @@ pub fn quantize(v: f32) -> i16 {
 // Link C AVX2 distance function when compiled with has_c_avx2
 #[cfg(has_c_avx2)]
 extern "C" {
-    fn l2sq_int16_avx2(a: *const i16, b: *const i16) -> i32;
+    fn l2sq_int16_avx2(a: *const i16, b: *const i16) -> i64;
 }
 
 #[cfg(test)]
